rock-paper-scissors3/challenge/requirements.txt
flask
requests
flask-csp
flask-wtf
flask-sqlalchemy
pyjwt